Ultra Basic stuff Started on
The Maze now technically exists (as a 10x10 matrix of "o") but at least
I got to the point where I remember how to make header files and link
them to the main.cpp again. I'll start on the actual implementation of
the nodes soon. I'm thinking it'll have to be nodes with flags for every
possible direction.